Cleaning the SPIDERS mess
