+++ app/frontend/scripts/main.js
@@ -6,10 +6,6 @@ import { sync } from 'vuex-router-sync'
 Vue.use(VeeValidate);
 Vue.use(VueRouter)
 
-const SOURCE_FILTER_NAME = 'source_';
-const INITIAL_FILTER_DESCRIPTIONS = CHOICES_FILTERS_TAGS.filter(c => c.name !== 'methods_')
-
-
 import store from './store/store.js';
 import routerGenerator from './routes/main.js'
 
@@ -26,10 +22,6 @@ document.addEventListener('DOMContentLoaded', () => {
     el: document.querySelector('#vue-content'),
     router,
     store,
-    // mounted: function () {
-    //   console.log("document.addEventListener / beforeCreate ... ")
-    //   store.dispatch('getConfigAll');
-    // },
     render: h => h( Vue.component('router-view') )
   })
